+++ app/controllers/admin/users_controller.rb
@@ -41,6 +41,24 @@ module Admin
 
     private
 
+    def respond_with_zipped_users
+      compressed_filestream = Zip::OutputStream.write_buffer do |zos|
+        User.order(created_at: :desc).each do |user|
+          zos.put_next_entry "user_#{user.id}.xlsx"
+          zos.print render_to_string(
+            layout: false,
+            handlers: [:axlsx],
+            formats: [:xlsx],
+            template: 'admin/users/user',
+            locals: { user: user }
+          )
+        end
+      end
+
+      compressed_filestream.rewind
+      send_data(compressed_filestream.read, filename: 'users.zip')
+    end
+
     def find_user
       @user = User.find(params[:id])
     end
@@ -51,22 +69,6 @@ module Admin
 
     def authorized_user!
       authorize(@user || User)
-    end
-
-    def respond_with_zipped_users
-      compressed_filestream = Zip::OutputStream.write_buffer do |zos|
-        User.order(created_at: :desc).each do |user|
-          zos.put_next_entry "user_#{user.id}.xlsx"
-          zos.print render_to_string(
-            layout: false, handlers: [:axlsx], formats: [:xlsx],
-            template: 'admin/users/user',
-            locals: { user: user }
-          )
-        end
-      end
-
-      compressed_filestream.rewind
-      send_data(compressed_filestream.read, filename: 'users.zip')
     end
   end
 end