+++ app/models/diet.rb
@@ -9,7 +9,6 @@ class Diet < ApplicationRecord
   validates :date, presence: true
 
   validate :correct_date
-  # validate :date_is_taken
 
   scope :search_by_day, ->(date) do
     diet = Diet.find_by(date: date || Date.today)
@@ -27,11 +26,4 @@ class Diet < ApplicationRecord
 
     errors.add(:date, 'is in the past.')
   end
-
-  # TODO: fix this and above validation
-  # def date_is_taken
-  #   return if date
-
-  #   errors.add(:date, 'is taken. You already have a diet for this date.') if Diet.find_by(date: date)
-  # end
 end